Capstone Project/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
sentence-transformers==2.7.0
faiss-cpu==1.8.0
streamlit==1.34.0
streamlit_chat==0.1.1
tqdm==4.66.4
PyPDF2==3.0.1
python-dotenv==1.0.0
altair==4
tiktoken==0.4.0
